pyproject.toml
[tool.poetry]
name = "etherscan-enot"
version = "0.1.0"
description = ""
authors = ["MrSterdy <mrsterdy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyppeteer = "^1.0.2"
asyncio = "^3.4.3"
tempmail-lol = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
